requirements.txt
python-dotenv
pandas
numpy
certifi
scikit-learn
dill
pyaml
mlflow
boto3
psycopg2-binary
# -e .